+++ src/main/groovy/com/monits/StaticCodeAnalysisPlugin.groovy
@@ -32,7 +32,7 @@ class StaticCodeAnalysisPlugin implements Plugin<Project> {
     private final static String BACKWARDS_PMD_TOOL_VERSION = '5.1.3'
     private final static GradleVersion GRADLE_VERSION_PMD = GradleVersion.version('2.4');
 
-    private final static String LATEST_CHECKSTYLE_VERSION = '6.17'
+    private final static String LATEST_CHECKSTYLE_VERSION = '6.18'
     private final static String BACKWARDS_CHECKSTYLE_VERSION = '6.7'
     private final static GradleVersion GRADLE_VERSION_CHECKSTYLE = GradleVersion.version('2.7');